add service for governance and other features (#253)
* add service for governance and other features
