+++ CMakeLists_Headers.txt
@@ -66,7 +66,6 @@ set(INCLUDE_HEADERS
 	src/manager/WFFuture.h
 	src/manager/WFFacilities.h
 	src/manager/WFFacilities.inl
-	src/manager/UpstreamPolicies.h
 	src/util/EncodeStream.h
 	src/util/LRUCache.h
 	src/util/StringUtil.h
@@ -83,8 +82,10 @@ set(INCLUDE_HEADERS
 	src/factory/WFAlgoTaskFactory.inl
 	src/factory/Workflow.h
 	src/factory/WFOperator.h
-	src/factory/WFNameService.h
-	src/factory/WFDNSResolver.h
+	src/nameservice/WFNameService.h
+	src/nameservice/WFDNSResolver.h
+	src/nameservice/ServiceGovernance.h
+	src/nameservice/UpstreamPolicies.h
 )
 
 if(KAFKA STREQUAL "y")
