+++ src/manager/CMakeLists.txt
@@ -6,7 +6,6 @@ set(SRC
 	UpstreamManager.cc
 	RouteManager.cc
 	WFGlobal.cc
-	UpstreamPolicies.cc
 )
 
 add_library(${PROJECT_NAME} OBJECT ${SRC})
