+++ src/nameservice/UpstreamPolicies.h
@@ -20,7 +20,6 @@
 #define _UPSTREAM_POLICIES_H_ 
 
 #include <pthread.h>
-#include <unordered_map>
 #include <vector>
 #include <atomic>
 #include "URIParser.h"
@@ -30,33 +29,23 @@
 #include "WFGlobal.h"
 #include "WFTaskError.h"
 #include "UpstreamManager.h"
-
-#define MTTR_SECOND			30
-#define VIRTUAL_GROUP_SIZE  16
-
-#define GET_CURRENT_SECOND  std::chrono::duration_cast<std::chrono::seconds>(std::chrono::steady_clock::now().time_since_epoch()).count()
+#include "ServiceGovernance.h"
 
 class EndpointGroup;
-class UPSPolicy;
 class UPSGroupPolicy;
 
-class EndpointAddress
+class UPSAddrParams : public PolicyAddrParams
 {
 public:
+	unsigned short weight;
+	short server_type;
+	int group_id;
 	EndpointGroup *group;
-	AddressParams params;
-	std::string address;
-	std::string host;
-	std::string port;
-	short port_value;
-	struct list_head list;
-	std::atomic<unsigned int> fail_count;
-	long long broken_timeout;
 	unsigned int consistent_hash[VIRTUAL_GROUP_SIZE];
 
-public:
-	EndpointAddress(const std::string& address,
-					const struct AddressParams *address_params);
+	UPSAddrParams();
+	UPSAddrParams(const struct AddressParams *params,
+				  const std::string& address);
 };
 
 class EndpointGroup
@@ -85,82 +74,18 @@ public:
 	const EndpointAddress *get_one_backup();
 };
 
-class UPSPolicy : public WFDNSResolver
-{
-public:
-	virtual WFRouterTask *create_router_task(const struct WFNSParams *params,
-											 router_callback_t callback);
-	virtual void success(RouteManager::RouteResult *result, void *cookie,
-					 	 CommTarget *target);
-	virtual void failed(RouteManager::RouteResult *result, void *cookie,
-						CommTarget *target);
-
-	void add_server(const std::string& address, const AddressParams *address_params);
-	int remove_server(const std::string& address);
-	int replace_server(const std::string& address, const AddressParams *address_params);
-
-	virtual void enable_server(const std::string& address);
-	virtual void disable_server(const std::string& address);
-	virtual void get_main_address(std::vector<std::string>& addr_list);
-	// virtual void server_list_change(/* std::vector<server> status */) {}
-
-public:
-	UPSPolicy() :
-		breaker_lock(PTHREAD_MUTEX_INITIALIZER),
-		rwlock(PTHREAD_RWLOCK_INITIALIZER)
-	{
-		this->nalives = 0;
-		this->try_another = false;
-		INIT_LIST_HEAD(&this->breaker_list);
-	}
-
-	virtual ~UPSPolicy()
-	{
-		for (EndpointAddress *addr : this->addresses)
-			delete addr;
-	}
-
-private:
-	virtual bool select(const ParsedURI& uri, EndpointAddress **addr);
-
-	virtual void recover_one_server(const EndpointAddress *addr)
-	{
-		this->nalives++;
-	}
-
-	virtual void fuse_one_server(const EndpointAddress *addr)
-	{
-		this->nalives--;
-	}
-
-	virtual void add_server_locked(EndpointAddress *addr);
-	virtual int remove_server_locked(const std::string& address);
-
-	void recover_server_from_breaker(EndpointAddress *addr);
-	void fuse_server_to_breaker(EndpointAddress *addr);
-
-	struct list_head breaker_list;
-	pthread_mutex_t breaker_lock;
-
-protected:
-	virtual const EndpointAddress *first_stradegy(const ParsedURI& uri);
-	virtual const EndpointAddress *another_stradegy(const ParsedURI& uri);
-	void check_breaker();
-
-	std::vector<EndpointAddress *> servers; // current servers
-	std::vector<EndpointAddress *> addresses; // memory management
-	std::unordered_map<std::string,
-					   std::vector<EndpointAddress *>> server_map;
-	pthread_rwlock_t rwlock;
-	std::atomic<int> nalives;
-	bool try_another;
-};
-
-class UPSGroupPolicy : public UPSPolicy
+class UPSGroupPolicy : public ServiceGovernance
 {
 public:
 	UPSGroupPolicy();
 	~UPSGroupPolicy();
+
+	virtual bool select(const ParsedURI& uri, EndpointAddress **addr);
+	virtual void add_server(const std::string& address,
+							const AddressParams *params);
+	virtual int replace_server(const std::string& address,
+							   const AddressParams *params);
+	void get_main_address(std::vector<std::string>& addr_list);
 
 protected:
 	struct rb_root group_map;
@@ -170,16 +95,16 @@ private:
 	virtual void recover_one_server(const EndpointAddress *addr)
 	{
 		this->nalives++;
-		addr->group->nalives++;
+		UPSAddrParams *params = static_cast<UPSAddrParams *>(addr->params);
+		params->group->nalives++;
 	}
 
 	virtual void fuse_one_server(const EndpointAddress *addr)
 	{
 		this->nalives--;
-		addr->group->nalives--;
+		UPSAddrParams *params = static_cast<UPSAddrParams *>(addr->params);
+		params->group->nalives--;
 	}
-
-	virtual bool select(const ParsedURI& uri, EndpointAddress **addr);
 
 protected:
 	virtual void add_server_locked(EndpointAddress *addr);
@@ -190,10 +115,11 @@ protected:
 
 	inline bool is_alive_or_group_alive(const EndpointAddress *addr) const
 	{
-		return ((addr->params.group_id < 0 &&
-					addr->fail_count < addr->params.max_fails) || 
-				(addr->params.group_id >= 0 &&
-					addr->group->nalives > 0));
+		UPSAddrParams *params = static_cast<UPSAddrParams *>(addr->params);
+		return ((params->group_id < 0 &&
+					addr->fail_count < addr->params->max_fails) ||
+				(params->group_id >= 0 &&
+					params->group->nalives > 0));
 	}
 };
