+++ src/nameservice/WFNameService.cc
